Rename code-defined skill registries for symmetry
