+++ front/lib/resources/skill/code_defined/discover_tools.ts
@@ -2,7 +2,7 @@ import { isJITMCPServerView } from "@app/lib/actions/mcp_internal_actions/utils"
 import { buildToolsetsContext } from "@app/lib/api/assistant/global_agents/configurations/dust/dust";
 import type { Authenticator } from "@app/lib/auth";
 import { MCPServerViewResource } from "@app/lib/resources/mcp_server_view_resource";
-import type { SystemSkillDefinition } from "@app/lib/resources/skill/global/shared";
+import type { SystemSkillDefinition } from "@app/lib/resources/skill/code_defined/shared";
 import type { AgentLoopExecutionData } from "@app/types/assistant/agent_run";
 
 export const discoverToolsSkill = {
