+++ front/lib/resources/skill/code_defined/shared.ts
