+++ front/tests/utils/SkillFactory.ts
@@ -1,8 +1,8 @@
 import type { Authenticator } from "@app/lib/auth";
 import { AgentSkillModel } from "@app/lib/models/agent/agent_skill";
 import type { MCPServerViewResource } from "@app/lib/resources/mcp_server_view_resource";
-import type { GlobalSkillId } from "@app/lib/resources/skill/global/registry";
-import type { SystemSkillId } from "@app/lib/resources/skill/global/system_registry";
+import type { GlobalSkillId } from "@app/lib/resources/skill/code_defined/global_registry";
+import type { SystemSkillId } from "@app/lib/resources/skill/code_defined/system_registry";
 import type { SkillAttachedKnowledge } from "@app/lib/resources/skill/skill_resource";
 import { SkillResource } from "@app/lib/resources/skill/skill_resource";
 import { SKILL_ICON } from "@app/lib/skill";
